==> 01_datacleaning.ipynb <==
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
from nltk.stem import WordNetLemmatizer, SnowballStemmer
from nltk.stem.porter import PorterStemmer
import re

train = pd.read_csv('./data/train.csv')
test = pd.read_csv('./data/test.csv')

# Lower casing all the text for some processing
# Removing the NA's from the keyword to filter and replace later
train['text'] = train['text'].str.lower()
train['keyword'] = train['keyword'].str.lower()
train['keyword'].fillna('-', inplace=True)

test['text'] = test['text'].str.lower()
test['keyword'] = test['keyword'].str.lower()
test['keyword'].fillna('-', inplace=True)

# This is removing all the URLs from the tweets.
# Basically all of them are t.co/ URLs which are shortened and 
# are mostly broken now.  

# Using this string remover from StackOverflow: 
# https://stackoverflow.com/questions/11331982/how-to-remove-any-url-within-a-string-in-python
train['text_nourl'] = [re.sub(r'\w+:\/{2}[\d\w-]+(\.[\d\w-]+)*(?:(?:\/[^\s/]*))*', '', i) for i in train['text']]
test['text_nourl'] = [re.sub(r'\w+:\/{2}[\d\w-]+(\.[\d\w-]+)*(?:(?:\/[^\s/]*))*', '', i) for i in test['text']]

# Removing the %20 from the keyword and making it a space
train['keyword'] = train['keyword'].str.replace('%20', ' ')
test['keyword'] = test['keyword'].str.replace('%20', ' ')

keywords_sorted = sorted(set(train['keyword']))

# Filling in most of the empty keys by using the existing keys 
# and checking which are in the tweet
def fill_keys(df):
    no_key = df.loc[df['keyword'] == '-']
    for index, row in no_key.iterrows():
        for word in keywords_sorted:
            if word in row['text']:
                df.loc[index, 'keyword'] = word
fill_keys(train)
fill_keys(test)

# Stemming keywords since many were similar
snowstem = SnowballStemmer('english')
train['keywords_stemmed'] = train['keyword'].map(lambda x: snowstem.stem(x))
test['keywords_stemmed'] = test['keyword'].map(lambda x: snowstem.stem(x))


train.head()

test.head()

train.to_csv('./data/train_clean.csv', index = False)
test.to_csv('./data/test_clean.csv', index = False)
